feat(MPDZBS-877 ZMS-1232): Refactor contoller tests and fix cors test
