+++ zmscitizenapi/tests/Zmscitizenapi/Controllers/Appointment/AppointmentPreconfirmTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace BO\Zmscitizenapi\Tests\Appointment;
+namespace BO\Zmscitizenapi\Tests\Controllers\Appointment;
 
 use BO\Zmscitizenapi\Localization\ErrorMessages;
-use BO\Zmscitizenapi\Tests\Base;
+use BO\Zmscitizenapi\Tests\ControllerTestCase;
 
-class AppointmentConfirmTest extends Base
+class AppointmentPreconfirmTest extends ControllerTestCase
 {
-    protected $classname = "\BO\Zmscitizenapi\Controllers\Appointment\AppointmentConfirmController";
+    protected $classname = "\BO\Zmscitizenapi\Controllers\Appointment\AppointmentPreconfirmController";
 
     public function setUp(): void
     {
@@ -42,13 +42,13 @@ class AppointmentConfirmTest extends Base
                 ],
                 [
                     'function' => 'readPostResult',
-                    'url' => '/process/status/confirmed/',
-                    'response' => $this->readFixture("POST_confirm_appointment.json")
+                    'url' => '/process/status/preconfirmed/',
+                    'response' => $this->readFixture("POST_preconfirm_appointment.json")
                 ],
                 [
                     'function' => 'readPostResult',
-                    'url' => '/process/101002/fb43/confirmation/mail/',
-                    'response' => $this->readFixture("POST_confirm_appointment.json")
+                    'url' => '/process/101002/fb43/preconfirmation/mail/',
+                    'response' => $this->readFixture("POST_preconfirm_appointment.json")
                 ]
             ]
         );
@@ -59,7 +59,7 @@ class AppointmentConfirmTest extends Base
         ];
         $response = $this->render([], $parameters, [], 'POST');
         $responseBody = json_decode((string) $response->getBody(), true);
-
+        
         $expectedResponse = [
             'processId' => 101002,
             'timestamp' => '1727865900',
@@ -85,7 +85,7 @@ class AppointmentConfirmTest extends Base
             'subRequestCounts' => [],
             'serviceId' => 10242339,
             'serviceCount' => 1,
-            'status' => 'confirmed'
+            'status' => 'preconfirmed'
         ];
     
         $this->assertEquals(200, $response->getStatusCode());
@@ -152,6 +152,49 @@ class AppointmentConfirmTest extends Base
         );
     }
 
+    public function testTooManyEmailsAtLocation()
+    {
+        $exception = new \BO\Zmsclient\Exception();
+        $exception->template = 'BO\\Zmsapi\\Exception\\Process\\MoreThanAllowedAppointmentsPerMail';
+
+        $this->setApiCalls([
+            [
+                'function' => 'readGetResult',
+                'url' => '/process/101002/fb43/',
+                'parameters' => [
+                    'resolveReferences' => 2,
+                ],
+                'response' => $this->readFixture("GET_process.json")
+            ],
+            [
+                'function' => 'readGetResult',
+                'url' => '/source/unittest/',
+                'parameters' => [
+                    'resolveReferences' => 2,
+                ],
+                'response' => $this->readFixture("GET_SourceGet_dldb.json")
+            ],
+            [
+                'function' => 'readPostResult',
+                'url' => '/process/status/preconfirmed/',
+                'exception' => $exception
+            ]
+        ]);
+
+        $parameters = [
+            'processId' => '101002',
+            'authKey' => 'fb43'
+        ];
+        $response = $this->render([], $parameters, [], 'POST');
+        $responseBody = json_decode((string) $response->getBody(), true);
+
+        $this->assertEquals(406, $response->getStatusCode());
+        $this->assertEqualsCanonicalizing(
+            ['errors' => [ErrorMessages::get('tooManyAppointmentsWithSameMail')]],
+            $responseBody
+        );
+    }
+
     public function testInvalidProcessId()
     {
         $parameters = [
@@ -214,11 +257,11 @@ class AppointmentConfirmTest extends Base
         );
     }
 
-    public function testNoEmailSendingWhenStatusNotConfirmed()
+    public function testNoEmailSendingWhenStatusNotPreconfirmed()
     {
-        $processResponse = $this->readFixture("POST_confirm_appointment.json");
+        $processResponse = $this->readFixture("POST_preconfirm_appointment.json");
         $processData = json_decode($processResponse, true);
-        $processData['data']['queue']['status'] = 'reserved'; // Change status to something else
+        $processData['data']['queue']['status'] = 'confirmed'; // Change status to something else
         
         $this->setApiCalls([
             [
@@ -239,7 +282,7 @@ class AppointmentConfirmTest extends Base
             ],
             [
                 'function' => 'readPostResult',
-                'url' => '/process/status/confirmed/',
+                'url' => '/process/status/preconfirmed/',
                 'response' => json_encode($processData)
             ]
             // Note: No email API call should be made
@@ -253,7 +296,7 @@ class AppointmentConfirmTest extends Base
         $responseBody = json_decode((string) $response->getBody(), true);
 
         $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('reserved', $responseBody['status']);
+        $this->assertEquals('confirmed', $responseBody['status']);
     }
 
     public function testInvalidRequest()
@@ -267,4 +310,47 @@ class AppointmentConfirmTest extends Base
             $responseBody
         );
     }
+
+    public function testPreconfirmationExpired()
+    {
+        $exception = new \BO\Zmsclient\Exception();
+        $exception->template = 'BO\\Zmsapi\\Exception\\Process\\PreconfirmationExpired';
+    
+        $this->setApiCalls([
+            [
+                'function' => 'readGetResult',
+                'url' => '/process/101002/fb43/',
+                'parameters' => [
+                    'resolveReferences' => 2,
+                ],
+                'response' => $this->readFixture("GET_process.json")
+            ],
+            [
+                'function' => 'readGetResult',
+                'url' => '/source/unittest/',
+                'parameters' => [
+                    'resolveReferences' => 2,
+                ],
+                'response' => $this->readFixture("GET_SourceGet_dldb.json")
+            ],
+            [
+                'function' => 'readPostResult',
+                'url' => '/process/status/preconfirmed/',
+                'exception' => $exception
+            ]
+        ]);
+    
+        $parameters = [
+            'processId' => '101002',
+            'authKey' => 'fb43'
+        ];
+        $response = $this->render([], $parameters, [], 'POST');
+        $responseBody = json_decode((string) $response->getBody(), true); 
+        $this->assertEquals(ErrorMessages::get('preconfirmationExpired')['statusCode'], $response->getStatusCode());
+        $this->assertEqualsCanonicalizing(
+            ['errors' => [ErrorMessages::get('preconfirmationExpired')]],
+            $responseBody
+        );
+    }
+
 }
