+++ zmscitizenapi/tests/Zmscitizenapi/Controllers/Service/ServiceListByOfficeTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace BO\Zmscitizenapi\Tests\Service;
+namespace BO\Zmscitizenapi\Tests\Controllers\Service;
 
 use BO\Zmscitizenapi\Localization\ErrorMessages;
-use BO\Zmscitizenapi\Tests\Base;
+use BO\Zmscitizenapi\Tests\ControllerTestCase;
 
-class ServiceListByOfficeTest extends Base
+class ServiceListByOfficeTest extends ControllerTestCase
 {
 
     protected $classname = "\BO\Zmscitizenapi\Controllers\Service\ServiceListByOfficeController";
